feat(columine): support the full flat Arrow type set
ArrowType grows from seven planes to twenty-three: every flat Arrow type
except the nested and dictionary-encoded families. Tags 0..6 are frozen
because they are baked into the shipped wasm and into persisted fixtures;
the sixteen new planes append from 7.

The plane table is now declared once, by the `arrow_planes!` macro, which
generates the enum and `ArrowType::ALL` together so the two cannot drift,
and `from_u8` is derived from `ALL` instead of restating twenty-three
numbers.

`PlaneKind` replaces the old `ColumnType`: one classification from which
storage sizing, Arrow buffer count, IPC buffer shape, producer range
checks and reader widening all derive. That is what collapses the
twenty-three-arm dispatches in the column store, the IPC writer, the
compact validator and both extractors into eight-arm matches on the kind
— and it is why adding a plane to an existing kind now needs no change
outside the table.

Storage is sized by the plane's actual per-row cost: `capacity * width`
for the fixed-width kinds, one bit per row for Bool, nothing at all for
Null, and the `min(max_data_bytes, capacity * 128)` estimate only for the
variable-width kinds, which are the only kinds whose per-row cost the
schema does not state. An i8 column now reserves one byte per row instead
of the 128 a variable-width column reserves.

`ColumnStorage` holds two byte buffers instead of five optional ones, so
no access needs a `die!` to assert that this shape really does have that
buffer, and `DynamicColumn` carries its whole `SignalSchemaField` instead
of a bare tag, so a column knows its own plane kind and cannot disagree
with itself about nullability.

Signed and unsigned are separate planes because the widening read
differs: 0xFFFFFFFF is -1 on the four-byte signed plane and 4294967295 on
the four-byte unsigned one, and no width check can tell those apart.
Date32/Time32 ride the four-byte signed plane and
Date64/Time64/Timestamp/Duration ride the eight-byte one;
`logical_type_matches` admits exactly those and rejects every other type
of the same width, including the Time32/Time64 units Arrow does not allow.

The two reserved metadata bytes become a `u16` type_param, which is the
FixedSizeBinary byte width and must be zero for every other plane. That
makes the four-byte physical descriptor total: a fixed-size-binary field
without a width is no longer representable. The width is bounded by
MAX_VALUE_BYTES / MAX_EVENTS_PER_BATCH = 256, derived so that no
fixed-width column can out-allocate one variable-width column's budget.

Nested and dictionary-encoded types stay out, and the enum doc says why:
the retained-metadata contract encodes at most three buffers per field,
and a type owning child fields or a dictionary body needs a different
layout contract rather than another enum arm.

+++ packages/columine/crates/columine-parsing/src/cross_format.rs
@@ -77,16 +77,16 @@ pub(crate) fn coerce(kind: ArrowType, value: &Logical) -> Result<Option<ColumnVa
     match (kind, value) {
         (ArrowType::Utf8, Logical::Text(text)) => Ok(Some(ColumnValue::Utf8(text.clone()))),
         (ArrowType::Int32, Logical::Int(value)) => i32::try_from(*value)
-            .map(|narrow| Some(ColumnValue::Int64(i64::from(narrow))))
+            .map(|narrow| Some(ColumnValue::Int(i64::from(narrow))))
             .map_err(|_| ()),
-        (ArrowType::Int64, Logical::Int(value)) => Ok(Some(ColumnValue::Int64(*value))),
+        (ArrowType::Int64, Logical::Int(value)) => Ok(Some(ColumnValue::Int(*value))),
         (ArrowType::Int64, Logical::Text(text)) => text
             .parse::<i64>()
             .or_else(|_| parse_iso8601_to_micros(text).map_err(|_| ()))
-            .map(|micros| Some(ColumnValue::Int64(micros)))
+            .map(|micros| Some(ColumnValue::Int(micros)))
             .map_err(|_| ()),
-        (ArrowType::Float64, Logical::Int(value)) => Ok(Some(ColumnValue::Float64(*value as f64))),
-        (ArrowType::Float64, Logical::Float(value)) => Ok(Some(ColumnValue::Float64(*value))),
+        (ArrowType::Float64, Logical::Int(value)) => Ok(Some(ColumnValue::Float(*value as f64))),
+        (ArrowType::Float64, Logical::Float(value)) => Ok(Some(ColumnValue::Float(*value))),
         (ArrowType::Bool, Logical::Bool(value)) => Ok(Some(ColumnValue::Bool(*value))),
         (ArrowType::Binary, value) => {
             Ok(Some(ColumnValue::Binary(canonical_msgpack(&value.json()))))
@@ -275,21 +275,21 @@ mod pins {
         // Int64 takes bigint-as-string on both wires.
         assert_eq!(
             via_json(ArrowType::Int64, "\"12345\""),
-            Ok(Some(ColumnValue::Int64(12345)))
+            Ok(Some(ColumnValue::Int(12345)))
         );
         assert_eq!(
             via_msgpack(ArrowType::Int64, "\"12345\""),
-            Ok(Some(ColumnValue::Int64(12345)))
+            Ok(Some(ColumnValue::Int(12345)))
         );
         // Int64 takes ISO-8601 on both wires.
         let micros = parse_iso8601_to_micros("2024-01-15T10:30:00Z").unwrap();
         assert_eq!(
             via_json(ArrowType::Int64, "\"2024-01-15T10:30:00Z\""),
-            Ok(Some(ColumnValue::Int64(micros)))
+            Ok(Some(ColumnValue::Int(micros)))
         );
         assert_eq!(
             via_msgpack(ArrowType::Int64, "\"2024-01-15T10:30:00Z\""),
-            Ok(Some(ColumnValue::Int64(micros)))
+            Ok(Some(ColumnValue::Int(micros)))
         );
         // Int64 takes no floats.
         assert_eq!(
@@ -308,15 +308,15 @@ mod pins {
     fn negative_int32_round_trips_signed() {
         assert_eq!(
             via_json(ArrowType::Int32, "-1"),
-            Ok(Some(ColumnValue::Int64(-1)))
+            Ok(Some(ColumnValue::Int(-1)))
         );
         assert_eq!(
             via_msgpack(ArrowType::Int32, "-1"),
-            Ok(Some(ColumnValue::Int64(-1)))
+            Ok(Some(ColumnValue::Int(-1)))
         );
         assert_eq!(
             via_json(ArrowType::Int32, "-2147483648"),
-            Ok(Some(ColumnValue::Int64(i64::from(i32::MIN))))
+            Ok(Some(ColumnValue::Int(i64::from(i32::MIN))))
         );
     }
 }
